+++ deepdoc/parser/excel_parser.py
@@ -6,7 +6,7 @@ from io import BytesIO
 from rag.nlp import find_codec
 
 
-class HuExcelParser:
+class RAGFlowExcelParser:
     def html(self, fnm):
         if isinstance(fnm, str):
             wb = load_workbook(fnm)
@@ -74,5 +74,5 @@ class HuExcelParser:
 
 
 if __name__ == "__main__":
-    psr = HuExcelParser()
+    psr = RAGFlowExcelParser()
     psr(sys.argv[1])
