+++ rag/nlp/synonym.py
@@ -17,12 +17,12 @@ class Dealer:
         try:
             self.dictionary = json.load(open(path, 'r'))
         except Exception as e:
-            logging.warning("Missing synonym.json")
+            logging.warn("Missing synonym.json")
             self.dictionary = {}
 
         if not redis:
             logging.warning(
-                "Real-time synonym is disabled, since no redis connection.")
+                "Realtime synonym is disabled, since no redis connection.")
         if not len(self.dictionary.keys()):
             logging.warning(f"Fail to load synonym")
