+++ rag/svr/cache_file_svr.py
@@ -4,7 +4,7 @@ import traceback
 
 from api.db.db_models import close_connection
 from api.db.services.task_service import TaskService
-from rag.utils import MINIO
+from rag.utils.minio_conn import MINIO
 from rag.utils.redis_conn import REDIS_CONN
